Move load_graph to file_utils from qc_utils
