+++ cat_merge/qc_utils.py
@@ -159,34 +159,6 @@ def get_difference(a: Union[List, pd.Series], b: Union[List, pd.Series]) -> Unio
     return s if type(a) is list else pd.Series(s, dtype=a.dtype, name=a.name)
 
 
-def load_graph(name: str, version: str, edges_path: str,
-               nodes_path: str) -> Graph:
-    """
-    Load a graph with Ensmallen (from grape).
-    :param name: OBO name
-    :param version: OBO version
-    :param edges_path: path to edge file
-    :param nodes_path: path to node file
-    :return: ensmallen Graph object
-    """
-
-    loaded_graph = Graph.from_csv(name=f"{name}_version_{version}",
-                                  edge_path=edges_path,
-                                  sources_column="subject",
-                                  destinations_column="object",
-                                  edge_list_header=True,
-                                  edge_list_separator="\t",
-                                  node_path=nodes_path,
-                                  nodes_column="id",
-                                  node_list_header=True,
-                                  node_list_separator="\t",
-                                  directed=False,
-                                  verbose=True
-                                  )
-
-    return loaded_graph
-
-
 def create_stats_report(g: Graph) -> List[Dict]:
     node_count = g.get_number_of_nodes()
     edge_count = g.get_number_of_edges()
